refactor(cli): migrate server/backend/config output to structured CommandResult
Phase 3.2 of the CLI output system refactor. Migrate 8 command handlers
from ad-hoc cliprint/fmt to clioutput.CommandResult + Renderer:

- backend.go: Full migration (handleBackendStatus, handleBackendSet)
- config.go: Partial migration (handleConfigSet, handleConfigList);
  handleConfigGet/Path kept as raw fmt.Println for piping
- server.go: Split from 896 lines into 5 focused files
- server_status.go: handleServerStatus via composable section-builders
- server_llm.go: addLLMSections section-builder, handleLLMList migrated
- server_health.go: Pure utility functions extracted

Introduces the section-builder pattern where functions like
addLLMSections(result, cfg) append sections to an existing CommandResult,
enabling reuse across standalone commands and embedded dashboards.

Deliberately excludes handleServerStart/handleLLMPull (ProgressDisplay
paradigm) and server_logs.go (log streaming).

Made-with: Cursor
