+++ client-apps/cli/cmd/stigmer/root/BUILD.bazel
@@ -8,6 +8,7 @@ go_library(
         "backend.go",
         "completion.go",
         "config.go",
+        "config_values.go",
         "delete.go",
         "delete_cancel.go",
         "delete_handlers.go",
@@ -46,7 +47,10 @@ go_library(
         "run_stream_subagent.go",
         "search.go",
         "server.go",
+        "server_health.go",
+        "server_llm.go",
         "server_logs.go",
+        "server_status.go",
         "validate.go",
         "verb_helpers.go",
     ],
